Add files via upload
Dask DataFrame
Dask DataFrame helps you process large tabular data by parallelizing pandas, either on your laptop for larger-than-memory computing, or on a distributed cluster of computers.

Just pandas: Dask DataFrames are a collection of many pandas DataFrames.

The API is the same. The execution is the same.

Large scale: Works on 100 GiB on a laptop, or 100 TiB on a cluster.

Easy to use: Pure Python, easy to set up and debug.